Php 8 upgrade (#17)
* First quick fixes to get it up and running

* Start using native gettext extension

* Update fpdf version to 1.86

* Fix deprecations and some null data issues

* Fix deprecations

* Fix deprecations and other issues

* Update the list of available locales

Since php depends on system gettext and locales available,
show only locales working.

* Remove facebook integration

* Delete twitter functionality

Twitter is now X, but ultiorganizer
is still ultiorganizer

* Remove svnclient

No reason to have this code in the project.

* More twitter removal

* Remove test code

* Fix deprecations also from comments

* Update readme

* Delete restful API

Implementation was very old and most probably
not even working.

* Proofreading some user facing content

* Small fixes

* Fix potential places for SQL injections

* Fix potential issues for cross site scripting

* Add validation to view loading

* Replace deprecated functionality

* Update session management to more secure

* Small fixes

* Hardening some function calls

* Update password hash

* Fix issues found during testing

* Fix php errors

* Update characterset to UTF8mb4

* Fix FFDA player management

* Get rid of utf8_encode functions

* Remove old jquery and related libs

It was not really used for anything useful.

* Update scorekeeper style

* Add check for utf8 conversion

* Database update

Change to innoDB and add foreign keys.

* Add user option to convert db tables to innodb

* Remove code not needed and prevent direct access to libs

* Make QR code bigger and fix pdf opening to own tab

* Update pdf names and opening method

* Add option to delete orphan player profiles

* Improve visitor count logging

* Add phpinfo view

* Add guards to prevent direct access to most pages without authentication

* Fix security issues

* Add info that PHP 8.3 is now supported

* Some grammar and spelling fixes

* Fix pdf showing on new page

* Hardening and fixing issues qith SQL queries

* Update translations

* Add undo option for season stats

* Fix stats calculation after foreign key restriction

* Add check to avoid page renders with invalid values

* Adjust qr-code printing on ffda score sheets

* Add parameter checks

* Handle invalid and missing parameters

* Minor fix to scorekeeper index

* Database query optimization for performance

* Small fixes to php errors seen from logs

* Streamline installation process

* Update SQL to match production version

---------

+++ admin/dbrestore.php
@@ -1,52 +1,51 @@
 <?php
+include_once __DIR__ . '/auth.php';
 include_once 'menufunctions.php';
 include_once 'lib/club.functions.php';
 include_once 'lib/reservation.functions.php';
 $html = "";
+if (!isSuperAdmin()) {
+	Forbidden(isset($_SESSION['uid']) ? $_SESSION['uid'] : 'anonymous');
+}
+
 if (!defined('ENABLE_ADMIN_DB_ACCESS') || constant('ENABLE_ADMIN_DB_ACCESS') != "enabled") {
 	$html = "<p>" . _("Direct database access is disabled. To enable it, define(ENABLE_ADMIN_DB_ACCESS,'enabled') in the config.inc.php file") . "</p>";
 } else {
 	if (isset($_POST['restore']) && isSuperAdmin()) {
-		if (is_uploaded_file($_FILES['restorefile']['tmp_name'])) {
+			if (is_uploaded_file($_FILES['restorefile']['tmp_name'])) {
+				$filenameParts = explode('.', $_FILES['restorefile']['name']);
+				$extension = strtolower(end($filenameParts));
 
-			if ("gz" == end(explode('.', $_FILES['restorefile']['name']))) {
-				$lines = gzfile($_FILES['restorefile']['tmp_name']);
-			} elseif ("sql" == end(explode('.', $_FILES['restorefile']['name']))) {
-				$lines = file($_FILES['restorefile']['tmp_name']);
-			}
+				if ("gz" == $extension) {
+					$lines = gzfile($_FILES['restorefile']['tmp_name']);
+				} elseif ("sql" == $extension) {
+					$lines = file($_FILES['restorefile']['tmp_name']);
+				} else {
+					$lines = null;
+					$html .= "<p>" . _("Unsupported backup format. Please upload a .gz or .sql file.") . "</p>";
+				}
 
-			$templine = '';
-			set_time_limit(300);
-
-			foreach ($lines as $line) {
-				// Skip it if it's a comment
-				if (substr($line, 0, 2) == '--' || $line == '')
-					continue;
-
-				$templine .= $line;
-				if (substr(trim($line), -1, 1) == ';') {
-					DBQuery($templine);
+				if (is_array($lines)) {
 					$templine = '';
+					set_time_limit(300);
+
+					foreach ($lines as $line) {
+						// Skip it if it's a comment
+						if (substr($line, 0, 2) == '--' || $line == '')
+							continue;
+
+						$templine .= $line;
+						if (substr(trim($line), -1, 1) == ';') {
+							DBQuery($templine);
+							$templine = '';
+						}
+					}
+					unlink($_FILES['restorefile']['tmp_name']);
+					unset($_SESSION['dbversion']);
+					$html .= "<p>" . _("Restore") . "</p>";
 				}
 			}
-			unlink($_FILES['restorefile']['tmp_name']);
-			unset($_SESSION['dbversion']);
-			$html .= "<p>" . _("Restore") . "</p>";
-		}
 
-		//disable facebook and twitter updates after restore to avoid false postings 
-		//(f.ex. if restored database is used for testing purpose)
-		$settings = array();
-
-		$setting = array();
-		$setting['name'] = "FacebookEnabled";
-		$setting['value'] = "false";
-		$settings[] = $setting;
-		$setting['name'] = "TwitterEnabled";
-		$setting['value'] = "false";
-		$settings[] = $setting;
-
-		SetServerConf($settings);
 	}
 
 	if (isSuperAdmin()) {
@@ -64,7 +63,7 @@ if (!defined('ENABLE_ADMIN_DB_ACCESS') || constant('ENABLE_ADMIN_DB_ACCESS') != 
 		$html .= "<input class='button' type='button' name='takaisin'  value='" . _("Return") . "' onclick=\"window.location.href='?view=admin/dbadmin'\"/></p>";
 		$html .= "</form>";
 	} else {
-		$html .= "<p>" . _("User credentials does not match") . "</p>\n";
+		$html .= "<p>" . _("User credentials do not match") . "</p>\n";
 	}
 }
 //common page
